agent: provide AgentFiles and ReportIter in rd-agent-intf
* Move AgentFiles which provides access to all the report interface files
  from resctl-demo into rd-agent-intf.

* Implement and expose ReportIter which walks the agent report files.

These will be used for resctl-bench to interact with the agent.
