+++ rd-agent-intf/Cargo.toml
@@ -15,6 +15,7 @@ clap = "2.33.0"
 chrono = { version = "0.4.10", features = ["serde"] }
 enum-iterator = "0.5.0"
 lazy_static = "1.4.0"
+libc = "^0.2"
 log = "0.4.8"
 serde = { version = "1.0.104", features = ["derive"] }
 util = { path = "../util" }
