+++ rd-agent-intf/src/lib.rs
@@ -1,5 +1,9 @@
 // Copyright (c) Facebook, Inc. and its affiliates.
+use libc;
+use log::error;
 use serde::{Deserialize, Serialize};
+use std::io;
+use util::*;
 
 pub mod args;
 pub mod bench;
@@ -20,7 +24,7 @@ pub use index::Index;
 pub use oomd::{OomdKnobs, OomdSliceMemPressureKnobs, OomdSliceSenpaiKnobs};
 pub use report::{
     BenchHashdReport, BenchIoCostReport, HashdReport, IoCostReport, IoLatReport, OomdReport,
-    Report, ResCtlReport, SideloadReport, SideloaderReport, SvcReport, SvcStateReport,
+    Report, ReportIter, ResCtlReport, SideloadReport, SideloaderReport, SvcReport, SvcStateReport,
     SysloadReport, UsageReport,
 };
 pub use side_defs::{SideloadDefs, SideloadSpec};
@@ -44,3 +48,86 @@ pub const OOMD_SVC_NAME: &str = "rd-oomd.service";
 pub const SIDELOADER_SVC_NAME: &str = "rd-sideloader.service";
 pub const SIDELOAD_SVC_PREFIX: &str = "rd-sideload-";
 pub const SYSLOAD_SVC_PREFIX: &str = "rd-sysload-";
+
+#[derive(Default)]
+pub struct AgentFiles {
+    pub args_path: String,
+    pub index_path: String,
+    pub args: JsonConfigFile<Args>,
+    pub index: JsonConfigFile<Index>,
+    pub cmd: JsonConfigFile<Cmd>,
+    pub cmd_ack: JsonConfigFile<CmdAck>,
+    pub sysreqs: JsonConfigFile<SysReqsReport>,
+    pub report: JsonConfigFile<Report>,
+    pub bench: JsonConfigFile<BenchKnobs>,
+    pub slices: JsonConfigFile<SliceKnobs>,
+    pub oomd: JsonConfigFile<OomdKnobs>,
+}
+
+impl AgentFiles {
+    pub fn new(dir: &str) -> Self {
+        Self {
+            args_path: dir.to_string() + "/args.json",
+            index_path: dir.to_string() + "/index.json",
+            ..Default::default()
+        }
+    }
+
+    fn refresh_one<T>(file: &mut JsonConfigFile<T>, path: &str) -> bool
+    where
+        T: JsonLoad + JsonSave,
+    {
+        match &file.path {
+            None => match JsonConfigFile::<T>::load(path) {
+                Ok(v) => {
+                    *file = v;
+                    true
+                }
+                Err(e) => {
+                    match e.downcast_ref::<io::Error>() {
+                        Some(e) if e.raw_os_error() == Some(libc::ENOENT) => (),
+                        _ => error!("Failed to read {:?} ({:?})", path, &e),
+                    }
+                    false
+                }
+            },
+            Some(_) => match file.maybe_reload() {
+                Ok(v) => v,
+                Err(e) => {
+                    match e.downcast_ref::<io::Error>() {
+                        Some(e) if e.raw_os_error() == Some(libc::ENOENT) => (),
+                        _ => error!("Failed to reload {:?} ({:?})", path, &e),
+                    }
+                    false
+                }
+            },
+        }
+    }
+
+    pub fn refresh(&mut self) {
+        Self::refresh_one(&mut self.args, &self.args_path);
+
+        if Self::refresh_one(&mut self.index, &self.index_path) {
+            self.cmd = Default::default();
+            self.cmd_ack = Default::default();
+            self.sysreqs = Default::default();
+            self.report = Default::default();
+            self.bench = Default::default();
+            self.slices = Default::default();
+            self.oomd = Default::default();
+        }
+        if let None = self.index.path {
+            return;
+        }
+
+        let index = &self.index.data;
+
+        Self::refresh_one(&mut self.cmd, &index.cmd);
+        Self::refresh_one(&mut self.cmd_ack, &index.cmd_ack);
+        Self::refresh_one(&mut self.sysreqs, &index.sysreqs);
+        Self::refresh_one(&mut self.report, &index.report);
+        Self::refresh_one(&mut self.bench, &index.bench);
+        Self::refresh_one(&mut self.slices, &index.slices);
+        Self::refresh_one(&mut self.oomd, &index.oomd);
+    }
+}
